Move rule upgrade to separate file
